+++ Controller/Payment/Callback/Complete.php
@@ -8,7 +8,6 @@ use Magento\Framework\Controller\ResultFactory;
 use Magento\Framework\Exception\PaymentException;
 use Magento\Framework\App\Request\InvalidRequestException;
 use Magento\Framework\App\RequestInterface;
-//use Magento\Sales\Api\OrderRepositoryInterface;
 use Nuvei\Checkout\Model\Payment;
 
 /**