+++ Model/Api/AiPayLink.php
@@ -2,11 +2,23 @@
 
 namespace Nuvei\Checkout\Model\Api;
 
+use Magento\Framework\Webapi\Rest\Request;
 use Magento\Framework\Exception\LocalizedException;
 use Nuvei\Checkout\Api\AiPayLinkInterface;
-use Nuvei\Checkout\Model\AbstractRequest;
+use Nuvei\Checkout\Model\Config;
+use Nuvei\Checkout\Model\ReaderWriter;
+use Magento\Quote\Model\QuoteFactory;
+use Magento\Store\Model\StoreManagerInterface;
+use Magento\Catalog\Api\ProductRepositoryInterface;
+use Magento\Customer\Model\Group;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Quote\Api\CartRepositoryInterface;
+use Magento\Quote\Model\QuoteIdMaskFactory;
+use Magento\Framework\UrlInterface;
 
 /**
+ * Create Quote and return the pay link to the sender.
+ * 
  * @author Nuvei
  */
 class AiPayLink implements AiPayLinkInterface
@@ -15,31 +27,69 @@ class AiPayLink implements AiPayLinkInterface
     private $requestFactory;
     private $moduleConfig;
     private $apiRequest;
-    
+    private $incomingParams;
+    private $quoteFactory;
+    private $storeManager;
+    private $productRepository;
+    private $quoteRepository;
+    private $quoteIdMaskFactory;
+    private $urlBuilder;
+
     public function __construct(
-        \Nuvei\Checkout\Model\Config $moduleConfig,
-        \Magento\Framework\Webapi\Rest\Request $apiRequest,
-        \Nuvei\Checkout\Model\ReaderWriter $readerWriter,
-        \Nuvei\Checkout\Model\Request\Factory $requestFactory
+        Config $moduleConfig,
+        Request $apiRequest,
+        ReaderWriter $readerWriter,
+        QuoteFactory $quoteFactory,
+        StoreManagerInterface $storeManager,
+        ProductRepositoryInterface $productRepository,
+        CartRepositoryInterface $quoteRepository,
+        QuoteIdMaskFactory $quoteIdMaskFactory,
+        UrlInterface $urlBuilder
     ) {
         $this->readerWriter         = $readerWriter;
         $this->moduleConfig         = $moduleConfig;
-        $this->requestFactory       = $requestFactory;
         $this->apiRequest           = $apiRequest;
+        $this->quoteFactory         = $quoteFactory;
+        $this->storeManager         = $storeManager;
+        $this->productRepository    = $productRepository;
+        $this->quoteRepository      = $quoteRepository;
+        $this->quoteIdMaskFactory   = $quoteIdMaskFactory;
+        $this->urlBuilder           = $urlBuilder;
     }
-    
+
     /**
      * Example of expected JSON structure:
+     *
      * {
-     *   "amount": 10.00,
-     *   "quantity": 1,
-     *   "currency": "EUR",
-     *   "description": "Nike Air Zoom",
-     *   "sku": "nike-air-zoom-42",
-     *   "customer_email": "john@example.com",
-     *   "success_url": "https://mystore.com/thank-you",
-     *   "failure_url": "https://mystore.com/error"
-     * }
+            "billing_address": {
+                "firstname": "first",
+                "lastname": "last",
+                "street": "tsarigradsko",
+                "city": "sofia",
+                "postcode": "1234",
+                "telephone": "1234567",
+                "region": "sofia",
+                "state": "state",
+                "country": "BG",
+                "email": "miroslavs@nuvei.com"
+            },
+            "shipping_address": {
+                "firstname": "first",
+                "lastname": "last",
+                "street": "tsarigradsko",
+                "city": "sofia",
+                "postcode": "1234",
+                "telephone": "1234567",
+                "region": "sofia",
+                "country": "BG"
+            },
+            "items": [
+                {
+                    "sku": "24-WB05",
+                    "quantity": 1
+                }
+            ]
+        }
      */
     public function generate()
     {
@@ -47,30 +97,246 @@ class AiPayLink implements AiPayLinkInterface
         if (!$this->moduleConfig->getConfigValue('active')) {
             $msg = 'Mudule is not active.';
             $this->readerWriter->createLog($msg);
-            
+
             throw new LocalizedException(__($msg));
         }
-        
-        $params = $this->apiRequest->getBodyParams();
-        
+
+        $this->incomingParams = $this->apiRequest->getBodyParams();
+
+        $this->readerWriter->createLog(
+            $this->incomingParams,
+            'The Incoming data.',
+            'DEBUG'
+        );
+
         // validate the parameters
-        $this->validateInputData($params);
-        
-        $request    = $this->requestFactory->create(AbstractRequest::GET_PAYMENT_LINK);
-        $response   = $request->setParams($params)->process();
-        
+        $this->validateInputData($this->incomingParams);
+
+        $payLinkData = $this->createQuote();
+
         // success
-        if (!empty($response['status']) && 'success' == strtolower($response['status'])) {
-           return [
-               "status"        => "success",
-               "paylink_url"   => $response['paymentPageUrl'],
-           ];
+        if (isset($payLinkData['responseCode']) && 200 == $payLinkData['responseCode']) {
+            http_response_code(200);
+            header('Content-Type: application/json');
+
+            exit(json_encode([
+                "status"        => "success",
+                "paylink_url"   => $payLinkData['payLink'],
+            ]));
         }
-        
-        // TODO - implement 4xx response code according to the error
-        return $response;
+
+        // error
+        http_response_code($payLinkData['responseCode']);
+
+        exit(json_encode([
+            "status"    => "error",
+            "message"   => $payLinkData['message'],
+        ]));
     }
-    
+
+    private function createQuote()
+    {
+        # Prepare quote
+        $store      = $this->storeManager->getStore();
+        $quote      = $this->quoteFactory->create();
+        $addedItems = 0;
+
+        $quote->setStore($store);
+        $quote->setCustomerIsGuest(true);
+        $quote->setCustomerGroupId(Group::NOT_LOGGED_IN_ID);
+
+        try {
+            $quote->setCustomerEmail($this->incomingParams['billing_address']['email']);
+            $quote->setIsActive(true);
+
+            // Load products by SKU and add them to the quote
+            foreach ($this->incomingParams['items'] as $itemData) {
+                $product        = $this->productRepository->get($itemData['sku']);
+                $buyRequestData = ['qty' => $itemData['quantity']];
+
+                // TODO - check if the product is with rebilling
+
+                switch ($product->getTypeId()) {
+                    case 'configurable':
+                        if (isset($itemData['super_attribute'])) {
+                            $buyRequestData['super_attribute'] = $itemData['super_attribute'];
+                        }
+                        break;
+
+                    case 'bundle':
+                        if (isset($itemData['bundle_option'])) {
+                            $buyRequestData['bundle_option']        = $itemData['bundle_option'];
+                            $buyRequestData['bundle_option_qty']    = $itemData['bundle_option_qty'] ?? [];
+                        }
+                        break;
+
+                    case 'grouped':
+                        if (isset($itemData['super_group'])) {
+                            $buyRequestData['super_group'] = $itemData['super_group'];
+                        }
+                        break;
+                }
+
+                $buyRequest = new \Magento\Framework\DataObject($buyRequestData);
+                $item       = $quote->addProduct($product, $buyRequest);
+
+                // error - the item is not an object but error message
+                if (is_string($item)) {
+                    $this->readerWriter->createLog(
+                        [
+                            'item data'     => $itemData,
+                            'error message' => $item,
+                        ],
+                        'This item was not added to the Quote.',
+                        'DEBUG'
+                    );
+
+                    continue;
+                }
+
+                $addedItems++;
+            }
+
+            // error - no added products in the quote
+            if (0 == $addedItems) {
+                $msg = __('The items were not added to the Cart. Abord the Order process.');
+
+                $this->readerWriter->createLog($e, $msg, 'DEBUG');
+
+                return [
+                    'message'       => $msg,
+                    'responseCode'  => 422,
+                ];
+            }
+
+            // Set billing address from incomingParams
+            $billingAddressData = [
+                'firstname'     => $this->incomingParams['billing_address']['firstname'] ?? '',
+                'lastname'      => $this->incomingParams['billing_address']['lastname'] ?? '',
+                'street'        => $this->incomingParams['billing_address']['street'] ?? '',
+                'city'          => $this->incomingParams['billing_address']['city'] ?? '',
+                'postcode'      => $this->incomingParams['billing_address']['postcode'] ?? '',
+                'country_id'    => $this->incomingParams['billing_address']['country'] ?? '',
+                'region'        => $this->incomingParams['billing_address']['region'] ?? '',
+                'telephone'     => $this->incomingParams['billing_address']['telephone'] ?? '',
+                'email'         => $this->incomingParams['billing_address']['email'],
+            ];
+            $billingAddress = $quote->getBillingAddress();
+            $billingAddress->addData($billingAddressData);
+
+            // Set shipping address (placeholder, replace with actual shipping address data later)
+            if (!$quote->isVirtual()) {
+                $shippingAddressData = [
+                    'firstname'     => $this->incomingParams['shipping_address']['firstname'] ?? '',
+                    'lastname'      => $this->incomingParams['shipping_address']['lastname'] ?? '',
+                    'street'        => $this->incomingParams['shipping_address']['street'] ?? '',
+                    'city'          => $this->incomingParams['shipping_address']['city'] ?? '',
+                    'postcode'      => $this->incomingParams['shipping_address']['postcode'] ?? '',
+                    'country_id'    => $this->incomingParams['shipping_address']['country'] ?? '',
+                    'region'        => $this->incomingParams['shipping_address']['region'] ?? '',
+                    'telephone'     => $this->incomingParams['shipping_address']['telephone'] ?? '',
+                ];
+                $shippingAddress = $quote->getShippingAddress();
+                $shippingAddress->addData($shippingAddressData);
+
+                // add shipping method
+                $shippingAddress->setCollectShippingRates(true)->collectShippingRates();
+
+                $rates              = $shippingAddress->getGroupedAllShippingRates();
+                $availableMethods   = [];
+                $selectedMethod     = '';
+
+                // collect the methods' codes
+                foreach ($rates as $carrierRates) {
+                    foreach ($carrierRates as $rate) {
+                        $availableMethods[] = $rate->getCode();
+                    }
+                }
+
+                // choose the method
+                if (in_array('freeshipping_freeshipping', $availableMethods)) {
+                    $selectedMethod = 'freeshipping_freeshipping';
+                }
+                elseif (in_array('flatrate_flatrate', $availableMethods)) {
+                    $selectedMethod = 'flatrate_flatrate';
+                }
+                else {
+                    $selectedMethod = $availableMethods[0];
+                }
+
+                $shippingAddress->setShippingMethod($selectedMethod);
+                $quote->collectTotals();
+            }
+
+            $quote->getPayment()->setMethod('nuvei');
+            $quote->collectTotals();
+            $this->quoteRepository->save($quote);
+
+            // build the payLink who will point back to the adapter
+            $quoteIdMask = $this->quoteIdMaskFactory->create()->load($quote->getId(), 'quote_id');
+            
+            if (!$quoteIdMask->getMaskedId()) {
+                $quoteIdMask->setQuoteId($quote->getId())->save();
+            }
+            
+            $maskedId   = $quoteIdMask->getMaskedId();
+            $payLink    = $this->urlBuilder->getUrl(
+                'nuvei_checkout/paybylink/redirect/',
+                ['quote' => $maskedId]
+            );
+
+            return [
+                'payLink'       => $payLink,
+                'responseCode'  => 200,
+            ];
+        }
+        catch (NoSuchEntityException $e) {
+            // Handle product not found
+            $msg = __('GetPaymentPageUrl Exception.');
+
+            $this->readerWriter->createLog(
+                [$e->getMessage()],
+                $msg,
+                'WARN'
+            );
+
+            return [
+                'message'       => $msg,
+                'responseCode'  => 500,
+            ];
+        }
+        catch (LocalizedException $e) {
+            // Handle Magento-specific errors
+            $msg = __('GetPaymentPageUrl LocalizedException.');
+
+            $this->readerWriter->createLog(
+                [$e->getMessage()],
+                $msg,
+                'WARN'
+            );
+
+            return [
+                'message'       => $msg,
+                'responseCode'  => 500,
+            ];
+        }
+        catch (\Exception $e) {
+            // Handle any other errors
+            $msg = __('GetPaymentPageUrl Exception.');
+
+            $this->readerWriter->createLog(
+                [$e->getMessage()],
+                $msg,
+                'WARN'
+            );
+
+            return [
+                'message'       => $msg,
+                'responseCode'  => 500,
+            ];
+        }
+    }
+
     /**
      * @param array $params
      * @return void
@@ -78,35 +344,35 @@ class AiPayLink implements AiPayLinkInterface
      */
     private function validateInputData($params)
     {
-        if (empty($params['amount']) || !is_numeric($params['amount'])) {
-            throw new LocalizedException(__('Invalid or missing amount.'));
+        if (empty($params['billing_address']) || !is_array($params['billing_address'])) {
+            throw new LocalizedException(__('Invalid or missing billing_address.'));
         }
-        if (empty($params['quantity']) || !is_int($params['quantity'])) {
-            throw new LocalizedException(__('Invalid or missing quantity.'));
-        }
-        if (empty($params['currency']) 
-            || !is_string($params['currency'])
-            || strlen($params['currency']) != 3
+        if (empty($params['billing_address']['email'])
+            || !filter_var($params['billing_address']['email'], FILTER_VALIDATE_EMAIL)
         ) {
-            throw new LocalizedException(__('Invalid or missing currency.'));
+            throw new LocalizedException(__('Invalid or missing billing_address - email.'));
         }
-        if (empty($params['description'])) {
-            throw new LocalizedException(__('Invalid or missing description.'));
+        if (empty($params['shipping_address']) || !is_array($params['shipping_address'])) {
+            throw new LocalizedException(__('Invalid or missing shipping_address.'));
         }
-        if (empty($params['sku'])) {
-            throw new LocalizedException(__('Invalid or missing sku.'));
+
+        // validate items
+        if (empty($params['items']) || !is_array($params['items'])) {
+            throw new LocalizedException(__('Invalid or missing items block.'));
         }
-        if (empty($params['customer_email']) || !filter_var($params['customer_email'], FILTER_VALIDATE_EMAIL)) {
-            throw new LocalizedException(__('Invalid or missing customer_email.'));
+        else {
+            foreach ($params['items'] as $itemData) {
+                if (empty($itemData['sku'])) {
+                    throw new LocalizedException(__('Invalid or missing item sku.'));
+                }
+                if (empty($itemData['quantity']) || !is_numeric($itemData['quantity'])) {
+                    throw new LocalizedException(__('Invalid or missing item quantity.'));
+                }
+            }
         }
-        if (empty($params['success_url']) || !filter_var($params['success_url'], FILTER_VALIDATE_URL)) {
-            throw new LocalizedException(__('Invalid or missing success_url.'));
-        }
-        if (empty($params['failure_url']) || !filter_var($params['failure_url'], FILTER_VALIDATE_URL)) {
-            throw new LocalizedException(__('Invalid or missing failure_url.'));
-        }
-        
+
+        // the date is valid
         return;
     }
-    
+
 }