+++ Model/Request/Factory.php
@@ -32,7 +32,6 @@ class Factory
         AbstractRequest::GET_SESSION_TOKEN                  => \Nuvei\Checkout\Model\Request\GetSessionToken::class,
         AbstractRequest::DELETE_UPOS_METHOD                 => \Nuvei\Checkout\Model\Request\DeleteUPO::class,
         AbstractRequest::PAYMENT_VOID_METHOD                => \Nuvei\Checkout\Model\Request\Payment\Cancel::class,
-        AbstractRequest::GET_PAYMENT_LINK                   => \Nuvei\Checkout\Model\Request\GetPaymentPageUrl::class,
     ];
 
     /**